Compile error on GCC.

+++ include/kitty/static_truth_table.hpp
@@ -139,7 +139,7 @@ struct static_truth_table<NumVars, false>
 
   enum
   {
-    NumBits = 1 << NumVars
+    NumBits = uint64_t( 1 ) << NumVars
   };
   /*! \endcond */
